backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
torch>=2.0.0
torchvision>=0.15.0
numpy==1.26.4
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.9.2
# Testing
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2


